connected search endpoint to subcategory page
