+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/ui/fragments/SubcategoryFragment.kt
@@ -8,22 +8,16 @@ import android.widget.AdapterView
 import android.widget.ArrayAdapter
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import androidx.navigation.Navigation
 import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.setupWithNavController
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.getflix.R
 import com.example.getflix.databinding.FragmentSubcategoryBinding
-import com.example.getflix.models.ProductModel
 import com.example.getflix.ui.adapters.SubCategoryAdapter
 import com.example.getflix.ui.fragments.SubcategoryFragmentDirections.Companion.actionSubcategoryFragmentToFilterOptionsFragment
 import com.example.getflix.ui.viewmodels.SubCategoryViewModel
 import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.activity_main.toolbar
 import kotlinx.android.synthetic.main.fragment_subcategory.*
 
 
@@ -41,51 +35,37 @@ class SubcategoryFragment : Fragment() {
                 container, false
         )
 
-        //activity?.toolbar!!.visibility = View.GONE
-        /* val navController = Navigation.findNavController(requireActivity(), R.id.my_nav_host_fragment)
-         val appBarConfiguration = AppBarConfiguration(navController.graph)
-         binding.toolbar.setupWithNavController(navController, appBarConfiguration)
-         binding.toolbar.title = ""
-         val args = SubcategoryFragmentArgs.fromBundle(requireArguments())
-         val subCat = args.subName
-         binding.toolbarTitle.text = subCat */
+        val args = SubcategoryFragmentArgs.fromBundle(requireArguments())
+        val subId = args.subId
         viewModel = ViewModelProvider(this).get(SubCategoryViewModel::class.java)
+        viewModel.searchBySubcategory(subId)
         //binding.viewmodel = SubCategoryViewModel()
         val recView = binding?.productList as RecyclerView
-        /*var zaraJacket1 =
-                ProductModel(10, "Jacket", "2", "1", "Zara", 1, 1, 1, "Nice jacket", "1", "1", "1", "1")
-        var zaraJacket2 =
-                ProductModel(11, "Jacket", "31", "1", "Zara", 1, 1, 1, "Cool jacket", "1", "1", "1", "1")
-        var zaraJacket3 =
-                ProductModel(12, "Jacket", "3", "1", "Zara", 1, 1, 1, "Amazing jacket", "1", "1", "1", "1")
-        var zaraSkirt1 =
-                ProductModel(4, "Skirt", "79", "1", "Zara", 1, 1, 1, "Nice skirt", "1", "1", "1", "1")
-        var zaraSkirt2 =
-                ProductModel(5, "Skirt", "93", "1", "Zara", 1, 1, 1, "Cool skirt", "1", "1", "1", "1")
-        var zaraSkirt3 =
-                ProductModel(6, "Skirt", "102", "1", "Zara", 1, 1, 1, "Amazing skirt", "1", "1", "1", "1")
-        val products = arrayListOf(zaraSkirt3, zaraJacket1, zaraSkirt2, zaraJacket2, zaraSkirt1, zaraJacket3)*/
-        val products = arrayListOf<ProductModel>()
+
         val manager = GridLayoutManager(activity, 2)
         recView.layoutManager = manager
-        val productListAdapter = SubCategoryAdapter(products)
-        recView.adapter = productListAdapter
-        recView.setHasFixedSize(true)
 
-        for (product in products) {
+        viewModel.productList.observe(viewLifecycleOwner, {
+            val productListAdapter = SubCategoryAdapter(it!!)
+            recView.adapter = productListAdapter
+            recView.setHasFixedSize(true)
+        })
+
+
+        /*for (product in products) {
             viewModel.addProduct(product)
-        }
+        }*/
 
         binding.btnFilter.setOnClickListener {
             filter = true
             view?.findNavController()!!.navigate(actionSubcategoryFragmentToFilterOptionsFragment())
         }
 
-        viewModel.productList.observe(viewLifecycleOwner, {
+        /*viewModel.productList.observe(viewLifecycleOwner, {
             it?.let {
                 productListAdapter.submitList(it)
             }
-        })
+        })*/
 
         val sorts = resources.getStringArray(R.array.SortBy)
         val spinner = binding.spinner
