+++ ecommerce/android/Getflix/app/src/main/java/com/example/getflix/ui/viewmodels/SubCategoryViewModel.kt
@@ -3,7 +3,14 @@ package com.example.getflix.ui.viewmodels
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.example.getflix.models.AddressModel
 import com.example.getflix.models.ProductModel
+import com.example.getflix.service.GetflixApi
+import com.example.getflix.service.requests.ProSearchBySubcategoryRequest
+import com.example.getflix.service.responses.ProSearchBySubcategoryResponse
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 class SubCategoryViewModel : ViewModel() {
 
@@ -41,5 +48,26 @@ class SubCategoryViewModel : ViewModel() {
     private fun sortPrice() {
         val sorted: MutableList<ProductModel> = _productList.value!!.sortedBy { it.price.toInt() } as MutableList<ProductModel>
         _productList.value = sorted
+    }
+
+    fun searchBySubcategory(subId: Int) {
+        GetflixApi.getflixApiService.searchProductsBySubcategory(ProSearchBySubcategoryRequest(subId))
+            .enqueue(object :
+                Callback<ProSearchBySubcategoryResponse> {
+                override fun onFailure(call: Call<ProSearchBySubcategoryResponse>, t: Throwable) {
+
+                }
+
+                override fun onResponse(
+                    call: Call<ProSearchBySubcategoryResponse>,
+                    response: Response<ProSearchBySubcategoryResponse>
+                ) {
+                    println(response.body().toString())
+                    println(response.code())
+                    _productList.value = response.body()!!.data.products as MutableList<ProductModel>
+
+                }
+            }
+            )
     }
 }
